Copy Cordova js files in copy and in update
